pyproject.toml
[project]
name = "hecto-deepfake-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pillow>=12.0.0",
    "timm>=1.0.22",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
